refactor: move some composables to the views where they're used
RISDEV-7420
