+++ frontend/src/views/amending-law/articles/editor/AmendingLawArticleEditor.view.vue
@@ -10,7 +10,6 @@ import { useArticle } from "@/views/amending-law/articles/editor/useArticle"
 import { useEidPathParameter } from "@/composables/useEidPathParameter"
 import { useDokumentExpressionEliPathParameter } from "@/composables/useDokumentExpressionEliPathParameter"
 import { useAknElementEidSelection } from "@/views/amending-law/articles/editor/useAknElementEidSelection"
-import { useModHighlightClasses } from "@/composables/useTimeBoundaryHighlightClasses"
 import { useModHighlightClasses } from "@/composables/useZeitgrenzenHighlightClasses"
 import { useNormRenderHtml } from "@/composables/useNormRender"
 import { useNormXml } from "@/composables/useNormXml"
