+++ frontend/src/views/amending-law/articles/editor/useMultiSelection.ts
